[Platinum V] Title: 등산 마니아, Time: 268 ms, Memory: 58144 KB, Score: 100 point -BaekjoonHub

+++ 백준/Platinum/20188. 등산 마니아/등산 마니아.cc
@@ -13,7 +13,7 @@ int makeTree(int root, int parent){
 	for (auto c : adj[root]){
 		if (c == parent) continue;
 		child[root].push_back(c);
-		int subSize = makeTree(c, root);
+		long long subSize = makeTree(c, root);
 		size += subSize;
 		ans += 1LL * subSize * (N - subSize) + subSize * (subSize - 1) / 2;
 	}
